StepFunctions

Documentation for StepFunctions.

StepFunctions.SortedDomainIterator

StepFunctions.StepFunction

StepFunctions.ValueSweepIterator

StepFunctions.to_minimal_stepfct_data

StepFunctions.SortedDomainIterator — Type

struct SortedDomainIterator{S,T}

An iterator that iterates over the sorted union of the domains of the step functions.

StepFunctions.StepFunction — Type

struct StepFunction{X<:Real,Y}

A step function is a piecewise constant function. It is represented by vectors xs and ys as follows: $f(t) = \begin{cases} ys[i+1] & \text{if } xs[i] \leq t < xs[i+1] \\ ys[1] & \text{if } t < xs[1] \\ ys[end] & \text{if } t \geq xs[end] \end{cases}$

In particular, we must have length(xs)+1 == length(ys). Furthermore, xs must be sorted and must not contain Inf.

StepFunctions.ValueSweepIterator — Type

struct ValueSweepIterator{X,Y,XS}

Iterates over the values of a step function at the points in a sorted iterator xs.

StepFunctions.to_minimal_stepfct_data — Method

function to_minimal_stepfct_data!(xs, y0, ys)

Given a vector of xs and a vector of ys